Refactored the "misc" package to documentation. I'll be adding the integrated help code into this package. Added layer.xml so it's easy to find the path's for new actions in the "XML Layer" file
