+++ manifest.mf
@@ -1,6 +1,7 @@
 Manifest-Version: 1.0
 AutoUpdate-Show-In-Client: true
 OpenIDE-Module: com.codeigniter.netbeans
+OpenIDE-Module-Layer: com/codeigniter/netbeans/layer.xml
 OpenIDE-Module-Localizing-Bundle: com/codeigniter/netbeans/Bundle.properties
 OpenIDE-Module-Specification-Version: 1.0
