+++ src/com/codeigniter/netbeans/documentation/logo.png
